TIP-221: Move FamilyFactory
