+++ spec/Pim/Bundle/TransformBundle/Transformer/FamilyTransformerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\TransformBundle\Transformer;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Factory\FamilyFactory;
+use Pim\Component\Catalog\Factory\FamilyFactory;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoTransformerInterface;
 use Pim\Bundle\TransformBundle\Transformer\EntityTransformerInterface;
 use Pim\Bundle\TransformBundle\Transformer\Guesser\GuesserInterface;
