+++ src/Pim/Bundle/CatalogBundle/Resources/config/factories.yml
@@ -1,5 +1,5 @@
 parameters:
-    pim_catalog.factory.family.class:                Pim\Bundle\CatalogBundle\Factory\FamilyFactory
+    pim_catalog.factory.family.class:                Pim\Component\Catalog\Factory\FamilyFactory
     pim_catalog.factory.attribute_requirement.class: Pim\Bundle\CatalogBundle\Factory\AttributeRequirementFactory
     pim_catalog.factory.metric.class:                Pim\Component\Catalog\Factory\MetricFactory
     pim_catalog.factory.group.class:                 Pim\Bundle\CatalogBundle\Factory\GroupFactory
