+++ src/Pim/Bundle/TransformBundle/Transformer/FamilyTransformer.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\TransformBundle\Transformer;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
-use Pim\Bundle\CatalogBundle\Factory\FamilyFactory;
+use Pim\Component\Catalog\Factory\FamilyFactory;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoTransformerInterface;
 use Pim\Bundle\TransformBundle\Transformer\Guesser\GuesserInterface;
 use Pim\Component\Catalog\Model\FamilyInterface;
